Put sane_morphism into it's own file
