+++ link-grammar/linkage/sane.h
@@ -1,0 +1,4 @@
+
+#include "link-includes.h"
+
+bool sane_linkage_morphism(Sentence, Linkage, Parse_Options);
